Diabetic or Non-Diabetic Predictor
This Python code is mainly to build a  ML model to classify a patient has diabetes or not. Here we are using RandomForestClassifier model for the classification purpose. The dataset used here is downloaded from Kaggle , ' Pima Indians Diabetes Database ' . We got an accuracy of 76% and is well tested on new data sets.